refactor: Directory structure and better naming;
