+++ flutter_frontend/lib/HomePage/home_page.dart
@@ -1,9 +1,9 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
-import 'package:flutter_frontend/repository/authentication_api.dart';
 import 'package:flutter_frontend/repository/authentication_repository.dart';
-import 'package:flutter_frontend/view/home_view.dart';
-import 'package:flutter_frontend/viewmodel/home_cubit.dart';
+import 'package:flutter_frontend/repository/user_repository.dart';
+import 'package:flutter_frontend/HomePage/home_view.dart';
+import 'package:flutter_frontend/HomePage/home_cubit.dart';
 
 class HomePage extends StatelessWidget {
   const HomePage({super.key});
@@ -12,8 +12,8 @@ class HomePage extends StatelessWidget {
   Widget build(BuildContext context) {
     return BlocProvider(
       create: (BuildContext context) => HomeCubit(
-        context.read<AuthenticationApi>(),
         context.read<AuthenticationRepository>(),
+        context.read<UserRepository>(),
       ),
       child: const HomeView(),
     );
