+++ flutter_frontend/lib/LibraryPage/html_image.dart
