+++ flutter_frontend/lib/LibraryPage/library_page.dart
@@ -1,9 +1,9 @@
 import 'package:flutter/widgets.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
-import 'package:flutter_frontend/repository/authentication_repository.dart';
-import 'package:flutter_frontend/repository/edition_api.dart';
-import 'package:flutter_frontend/view/library_view.dart';
-import 'package:flutter_frontend/viewmodel/library_cubit.dart';
+import 'package:flutter_frontend/repository/user_repository.dart';
+import 'package:flutter_frontend/repository/edition_repository.dart';
+import 'package:flutter_frontend/LibraryPage/library_view.dart';
+import 'package:flutter_frontend/LibraryPage/library_cubit.dart';
 
 class LibraryPage extends StatelessWidget {
   const LibraryPage({required this.librarianId, super.key});
@@ -13,8 +13,8 @@ class LibraryPage extends StatelessWidget {
   Widget build(BuildContext context) {
     return BlocProvider(
       create: (BuildContext context) => LibraryCubit(
-        context.read<EditionApi>(),
-        context.read<AuthenticationRepository>(),
+        context.read<EditionRepository>(),
+        context.read<UserRepository>(),
         librarianId,
       )..getEditions(),
       child: const LibraryView(),
