+++ flutter_frontend/lib/LibraryPage/library_view.dart
@@ -1,8 +1,8 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
-import 'package:flutter_frontend/model/library_state.dart';
-import 'package:flutter_frontend/view/edition_view.dart';
-import 'package:flutter_frontend/viewmodel/library_cubit.dart';
+import 'package:flutter_frontend/LibraryPage/library_state.dart';
+import 'package:flutter_frontend/LibraryPage/edition_view.dart';
+import 'package:flutter_frontend/LibraryPage/library_cubit.dart';
 
 class LibraryView extends StatelessWidget {
   const LibraryView({super.key});
