+++ flutter_frontend/lib/main.dart
@@ -1,10 +1,10 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
-import 'package:flutter_frontend/repository/authentication_api.dart';
 import 'package:flutter_frontend/repository/authentication_repository.dart';
-import 'package:flutter_frontend/repository/edition_api.dart';
-import 'package:flutter_frontend/view/home_page.dart';
-import 'package:flutter_frontend/view/library_page.dart';
+import 'package:flutter_frontend/repository/user_repository.dart';
+import 'package:flutter_frontend/repository/edition_repository.dart';
+import 'package:flutter_frontend/HomePage/home_page.dart';
+import 'package:flutter_frontend/LibraryPage/library_page.dart';
 import 'package:go_router/go_router.dart';
 
 // GoRouter configuration
@@ -31,10 +31,10 @@ void main() async {
   runApp(
     MultiRepositoryProvider(
       providers: [
-        RepositoryProvider(create: (context) => EditionApi()),
-        RepositoryProvider(create: (context) => AuthenticationApi()),
+        RepositoryProvider(create: (context) => EditionRepository()),
+        RepositoryProvider(create: (context) => AuthenticationRepository()),
         RepositoryProvider(
-          create: (context) => AuthenticationRepository(),
+          create: (context) => UserRepository(),
         ),
       ],
       child: const MyApp(),
