--- flutter_frontend/lib/repository/authentication_api.dart
@@ -1,13 +1,0 @@
-import 'package:flutter_frontend/core/api.dart';
-
-class AuthenticationApi {
-  Future<bool> authenticateUser(String idToken) async {
-    final body = await get('/api/authenticate', {'idToken': idToken});
-
-    if (body == 'true') {
-      return true;
-    } else {
-      return false;
-    }
-  }
-}
