+++ flutter_frontend/lib/repository/authentication_repository.dart
@@ -1,20 +1,13 @@
-import 'package:flutter_frontend/model/user.dart';
+import 'package:flutter_frontend/core/api.dart';
 
 class AuthenticationRepository {
-  User? _cachedUser;
+  Future<bool> authenticateUser(String idToken) async {
+    final body = await get('/api/authenticate', {'idToken': idToken});
 
-  Future<User?> getCurrentUser() async {
-    if (_cachedUser != null) return _cachedUser;
-    return null;
-  }
-
-  Future<User> login(String email, String idToken) async {
-    final user = User(email: email, idToken: idToken);
-    _cachedUser = user;
-    return user;
-  }
-
-  Future<void> logout() async {
-    _cachedUser = null;
+    if (body == 'true') {
+      return true;
+    } else {
+      return false;
+    }
   }
 }
