+++ flutter_frontend/lib/repository/user_repository.dart
@@ -1,0 +1,20 @@
+import 'package:flutter_frontend/model/user.dart';
+
+class UserRepository {
+  User? _cachedUser;
+
+  Future<User?> getCurrentUser() async {
+    if (_cachedUser != null) return _cachedUser;
+    return null;
+  }
+
+  Future<User> login(String email, String idToken) async {
+    final user = User(email: email, idToken: idToken);
+    _cachedUser = user;
+    return user;
+  }
+
+  Future<void> logout() async {
+    _cachedUser = null;
+  }
+}
