--- flutter_frontend/lib/sign_in/google_sign_in.dart
@@ -1,30 +1,0 @@
-// import 'package:flutter/material.dart';
-// import 'package:flutter_bloc/flutter_bloc.dart';
-// import 'package:flutter_frontend/core/user_state.dart';
-// import 'package:flutter_frontend/main.dart';
-// import 'package:google_sign_in/google_sign_in.dart';
-// import 'package:google_sign_in_platform_interface/google_sign_in_platform_interface.dart';
-// import 'package:google_sign_in_web/google_sign_in_web.dart' as web;
-
-// class GoogleSignInPage extends StatefulWidget {
-//   const GoogleSignInPage({super.key});
-
-//   @override
-//   State<GoogleSignInPage> createState() => _GoogleSignInPageState();
-// }
-
-// class _GoogleSignInPageState extends State<GoogleSignInPage> {
-//   final GoogleSignInAccount? user = getIt<UserBloc>().state.user;
-
-//   @override
-//   Widget build(BuildContext context) {
-//     if (user == null) {
-//       return (GoogleSignInPlatform.instance as web.GoogleSignInPlugin)
-//           .renderButton();
-//     }
-
-//     return Center(
-//       child: GoogleUserCircleAvatar(identity: user!),
-//     );
-//   }
-// }
